requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
underthesea>=1.3.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pydantic>=1.8.0
python-dotenv>=0.19.0 